Custom Offline Payments: Move 'PaymentMethod' enum values to 'StaticPaymentMethods'
This change is done in preparation for user defined payment methods.
Previously, 'PaymentMethod' was an enum which would not easily
facilitate new user defined methods. The enum has been moved to
'StaticPaymentMethods' which implements the 'PaymentMethod' interface.
This interface can be used in the future to create other concrete
payment method types.