Custom Offline Payments - Move all COP DB queries into manager class
Before this change we were making a lot of arbitrary requests to read &
write the various data related to custom payments. This was brittle as
something like a JSON error from a single place could affect other parts
of the code. All of the reading and writing to the DB is now handled
through CustomOfflineConfigurationManager. This will make it much easier
in the future to change the schemas associated with the feature or to
make any other changes to how custom payment data is stored.